Fix #5 Support for multiple persistence units and DAOs
Apache Onami persist already provides support for multiple persistence units, but makes it difficult to provide generic DAOs linked to the annotations which identify the different PUs.  This is because the binding to annotations is done in a PrivateModule.  The Onami notes.md file provides a brief description of the changes made to the original code.

  The implementations provided by this change enable the use of multiple PUs, and provide generic DAOs to support them.  The base of an entity specific DAO is also provided