Fix #5 Fix#1 Implementation of DAOs and persistence for Option
onami-persist uses Guice Private Modules to enable the use of multiple persistence units.  However, many of its classes are inaccessible for modification, and this prevented binding DAOs through the same mechanism (which would be required of course to ensure that DAOs were pointing at the correct source).  This would not be an issue for Entity-specific DAOs, but precluded the use of generic DOAs.  Onami-persist has therefore been imported as source and modified slightly.  (see Onami notes.md for some detail).

Option is now persistable